refactor: 각 페이지들의 styled 하나로 통일
#11
