--- src/page/pokedex/Pokedex.styles.ts
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-export const PokedexContainer = styled.div`
-  max-width: 1200px;
-  margin: auto;
-  padding: 30px;
-  > *:not(:last-child) {
-    margin-bottom: 30px;
-  }
-`;
